requirements.txt
glfw
PyOpenGL
Pillow
PyGLM
freetype-py